add conftest.py; merge _create_report methods
