+++ checkov/common/bridgecrew/vulnerability_scanning/image_scanner.py
@@ -79,9 +79,12 @@ class ImageScanner:
         try:
             self.setup_scan(docker_image_id, dockerfile_path, skip_extract_image_name)
             scan_result = self.run_image_scan(docker_image_id)
-            docker_image_scanning_integration.report_results(self.docker_image_name, dockerfile_path,
-                                                             self.dockerfile_content,
-                                                             twistcli_scan_result=scan_result)
+            docker_image_scanning_integration.report_results(
+                twistcli_scan_result=scan_result,
+                file_path=dockerfile_path,
+                file_content=self.dockerfile_content,
+                docker_image_name=self.docker_image_name,
+            )
             logging.info(f'Docker image scanning results reported to the platform')
             self.cleanup_scan()
         except Exception as e:
