+++ checkov/common/bridgecrew/vulnerability_scanning/integrations/docker_image_scanning.py
@@ -11,24 +11,25 @@ from checkov.common.util.str_utils import removeprefix
 class DockerImageScanningIntegration(TwistcliIntegration):
     def report_results(
         self,
-        docker_image_name: str,
-        dockerfile_path: Path,
-        dockerfile_content: str,
         twistcli_scan_result: Dict[str, Any],
+        file_path: Path,
+        file_content: str,
+        docker_image_name: str,
+        **kwargs: Any,
     ) -> None:
         super().report_results(
-            docker_image_name=docker_image_name,
-            dockerfile_path=dockerfile_path,
-            dockerfile_content=dockerfile_content,
             twistcli_scan_result=twistcli_scan_result,
+            file_path=file_path,
+            file_content=file_content,
+            docker_image_name=docker_image_name,
         )
 
     def _create_report(
         self,
-        docker_image_name: str,
-        dockerfile_path: Path,
-        dockerfile_content: str,
         twistcli_scan_result: Dict[str, Any],
+        file_path: Path,
+        file_content: str,
+        docker_image_name: str,
     ) -> Dict[str, Any]:
         vulnerabilities = [
             {
@@ -45,13 +46,12 @@ class DockerImageScanningIntegration(TwistcliIntegration):
                 "publishedDate": vul.get("publishedDate")
                 or (datetime.now() - timedelta(days=vul.get("publishedDays", 0))).isoformat(),
             }
-            for vul in twistcli_scan_result.get("vulnerabilities") or []
+            for vul in twistcli_scan_result["results"][0].get("vulnerabilities") or []
         ]
         payload = {
             "dockerImageName": docker_image_name,
-            # BC_ROOT_DIR is used in the platform
-            "dockerFilePath": removeprefix(str(dockerfile_path), os.getenv("BC_ROOT_DIR", "")),
-            "dockerFileContent": dockerfile_content,
+            "dockerFilePath": removeprefix(str(file_path), os.getenv("BC_ROOT_DIR", "")),
+            "dockerFileContent": file_content,
             "type": "Image",
             "sourceId": bc_integration.repo_id,
             "branch": bc_integration.repo_branch,
@@ -59,16 +59,6 @@ class DockerImageScanningIntegration(TwistcliIntegration):
             "vulnerabilities": vulnerabilities,
         }
         return payload
-
-    # not used for image scanning
-    def _create_report_async(
-        self,
-        twistcli_scan_result: Dict[str, Any],
-        bc_platform_integration: BcPlatformIntegration,
-        file_path: Path,
-        **kwargs: Any,
-    ) -> Dict[str, Any]:
-        pass
 
 
 docker_image_scanning_integration = DockerImageScanningIntegration()
