+++ checkov/common/bridgecrew/vulnerability_scanning/integrations/package_scanning.py
@@ -9,11 +9,7 @@ from checkov.common.util.str_utils import removeprefix
 
 
 class PackageScanningIntegration(TwistcliIntegration):
-    # not used for package scanning
-    def _create_report(self, **kwargs: Any) -> Dict[str, Any]:
-        pass
-
-    def _create_report_async(
+    def _create_report(
         self,
         twistcli_scan_result: Dict[str, Any],
         bc_platform_integration: BcPlatformIntegration,
@@ -39,7 +35,6 @@ class PackageScanningIntegration(TwistcliIntegration):
         ]
         payload = {
             "packageName": str(file_path.name),
-            # BC_ROOT_DIR is used in the platform
             "packageFilePath": removeprefix(str(file_path), os.getenv("BC_ROOT_DIR", "")),
             "type": "Package",
             "sourceId": bc_platform_integration.repo_id,
