+++ checkov/common/bridgecrew/vulnerability_scanning/integrations/twistcli.py
@@ -40,19 +40,18 @@ class TwistcliIntegration(ABC):
         logging.debug(f"twistcli downloaded and has execute permission")
 
     # can be removed, if image scanning is also using asyncio
-    def report_results(self, **kwargs: Any) -> None:
-        payload = self._create_report(**kwargs)
+    def report_results(self, twistcli_scan_result: Dict[str, Any], file_path: Path, **kwargs: Any) -> None:
+        payload = self._create_report(
+            twistcli_scan_result=twistcli_scan_result,
+            file_path=file_path,
+            **kwargs,
+        )
         headers = merge_dicts(
             get_default_post_headers(bc_integration.bc_source, bc_integration.bc_source_version),
             {"Authorization": self.get_bc_api_key()},
         )
         response = requests.request("POST", f"{self.vulnerabilities_base_url}/report", headers=headers, json=payload)
         response.raise_for_status()
-
-    # can be removed, if image scanning is also using asyncio
-    @abstractmethod
-    def _create_report(self, **kwargs: Any) -> Dict[str, Any]:
-        pass
 
     async def report_results_async(
         self,
@@ -64,7 +63,7 @@ class TwistcliIntegration(ABC):
     ) -> int:
         logging.info(f"Start to send report for package file {file_path}")
 
-        payload = self._create_report_async(
+        payload = self._create_report(
             twistcli_scan_result=twistcli_scan_result,
             bc_platform_integration=bc_platform_integration,
             file_path=file_path,
@@ -90,11 +89,5 @@ class TwistcliIntegration(ABC):
                 return 1
 
     @abstractmethod
-    def _create_report_async(
-        self,
-        twistcli_scan_result: Dict[str, Any],
-        bc_platform_integration: BcPlatformIntegration,
-        file_path: Path,
-        **kwargs: Any,
-    ) -> Dict[str, Any]:
+    def _create_report(self, **kwargs: Any) -> Dict[str, Any]:
         pass
