+++ tests/common/bridgecrew/vulnerability_scanning/conftest.py
@@ -1,0 +1,51 @@
+from typing import Dict, Any
+
+import pytest
+
+from checkov.common.bridgecrew.bc_source import SourceType
+from checkov.common.bridgecrew.platform_integration import BcPlatformIntegration, bc_integration
+
+
+@pytest.fixture()
+def mock_bc_integration() -> BcPlatformIntegration:
+    bc_integration.bc_api_key = "abcd1234-abcd-1234-abcd-1234abcd1234"
+    bc_integration.setup_bridgecrew_credentials(
+        repo_id="bridgecrewio/checkov",
+        skip_fixes=True,
+        skip_suppressions=True,
+        skip_policy_download=True,
+        source=SourceType("Github", False),
+        source_version="1.0",
+        repo_branch="master",
+    )
+    return bc_integration
+
+
+@pytest.fixture()
+def scan_result() -> Dict[str, Any]:
+    return {
+        "repository": "/abs_path/to/app/requirements.txt",
+        "passed": True,
+        "packages": {"type": "python", "name": "django", "version": "1.2", "path": "/abs_path/to/app/requirements.txt"},
+        "complianceIssues": None,
+        "complianceDistribution": {"critical": 0, "high": 0, "medium": 0, "low": 0, "total": 0},
+        "vulnerabilities": [
+            {
+                "id": "CVE-2019-19844",
+                "status": "fixed in 3.0.1, 2.2.9, 1.11.27",
+                "cvss": 9.8,
+                "vector": "CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+                "description": "Django before 1.11.27, 2.x before 2.2.9, and 3.x before 3.0.1 allows account takeover.",
+                "severity": "critical",
+                "packageName": "django",
+                "packageVersion": "1.2",
+                "link": "https://nvd.nist.gov/vuln/detail/CVE-2019-19844",
+                "riskFactors": ["Critical severity", "Has fix", "Attack complexity: low", "Attack vector: network"],
+                "impactedVersions": ["\u003c1.11.27"],
+                "publishedDate": "2019-12-18T20:15:00+01:00",
+                "discoveredDate": "2019-12-18T19:15:00Z",
+                "fixDate": "2019-12-18T20:15:00+01:00",
+            }
+        ],
+        "vulnerabilityDistribution": {"critical": 1, "high": 0, "medium": 0, "low": 0, "total": 0},
+    }
